Add asset_type schema, helper, and refactor code
