+++ apps/ello_v3/lib/ello_v3/schema.ex
@@ -6,6 +6,7 @@ defmodule Ello.V3.Schema do
   import_types Absinthe.Type.Custom
   import_types __MODULE__.ContentTypes
   import_types __MODULE__.NetworkTypes
+  import_types __MODULE__.AssetTypes
 
   query do
     @desc "Get a post by username and token"
